Also for RAM and GPU.

+++ app/src/main/java/uk/co/syski/client/android/data/entity/RAMEntity.java
@@ -17,6 +17,6 @@ public class RAMEntity {
 
     public String MemoryTypeName;
 
-    public int MemoryBytes;
+    public long MemoryBytes;
 }
